Added graph
