+++ client/src/AuthWrapper.tsx
@@ -22,9 +22,8 @@ export const AuthWrapper = ({ children }: AuthContextProviderProps) => {
           return;
         }
         const userDetails = session?.user.user_metadata;
-        BackendAxios.defaults.headers.common[
-          "Authorization"
-        ] = `Bearer ${session?.access_token}`;
+        BackendAxios.defaults.headers.common["Authorization"] =
+          `Bearer ${session?.access_token}`;
 
         signIn({
           id: session?.user.id ?? "",
@@ -45,9 +44,8 @@ export const AuthWrapper = ({ children }: AuthContextProviderProps) => {
     } = supabaseClient.auth.onAuthStateChange((authEvent, session) => {
       switch (authEvent) {
         case "SIGNED_IN": {
-          BackendAxios.defaults.headers.common[
-            "Authorization"
-          ] = `Bearer ${session?.access_token}`;
+          BackendAxios.defaults.headers.common["Authorization"] =
+            `Bearer ${session?.access_token}`;
           setLoading(false);
 
           const userDetails = session?.user.user_metadata;
